Renamed the oracles to be consistent with current style.
