+++ eqtests/basic-eqtests/src/main/java/de/learnlib/eqtests/basic/RandomWMethodEQOracle.java
@@ -1,11 +1,10 @@
-package learnlib;
+package de.learnlib.eqtests.basic;
 
 import de.learnlib.api.EquivalenceOracle;
 import de.learnlib.api.MembershipOracle;
 import de.learnlib.oracles.DefaultQuery;
 import net.automatalib.automata.UniversalDeterministicAutomaton;
 import net.automatalib.automata.concepts.Output;
-import net.automatalib.commons.util.mappings.MutableMapping;
 import net.automatalib.util.automata.Automata;
 import net.automatalib.words.Word;
 import net.automatalib.words.WordBuilder;
@@ -14,12 +13,12 @@ import javax.annotation.ParametersAreNonnullByDefault;
 import java.util.*;
 
 /**
- * Implements an equivalence test by applying the Wp-method test on the given hypothesis automaton,
- * as described in "Test Selection Based on Finite State Models" by S. Fujiwara et al.
+ * Implements an equivalence test by applying the W-method test on the given hypothesis automaton.
+ * Generally the Wp-method performs better in finding counter examples.
  * Instead of enumerating the test suite in order, this is a sampling implementation:
  * 1. sample uniformly from the states for a prefix
  * 2. sample geometrically a random word
- * 3. sample a word from the set of suffixes / state identifiers (either local or global)
+ * 3. sample a word from the set of suffixes / state identifiers
  * There are two parameters: minimalSize determines the minimal size of the random word, this is
  * useful when one first performs a W(p)-method with some depth and continue with this randomized
  * tester from that depth onward. The second parameter rndLength determines the expected length
@@ -31,7 +30,7 @@ import java.util.*;
  * @param <D> output domain type
  * @author Joshua Moerman
  */
-public class RandomWpMethod<A extends UniversalDeterministicAutomaton<?, I, ?, ?, ?> & Output<I, D>, I, D>
+public class RandomWMethodEQOracle<A extends UniversalDeterministicAutomaton<?, I, ?, ?, ?> & Output<I, D>, I, D>
         implements EquivalenceOracle<A, I, D> {
     private final MembershipOracle<I, D> sulOracle;
     private final int minimalSize;
@@ -45,7 +44,7 @@ public class RandomWpMethod<A extends UniversalDeterministicAutomaton<?, I, ?, ?
      * @param minimalSize minimal size of the random word
      * @param rndLength   expected length (in addition to minimalSize) of random word
      */
-    public RandomWpMethod(MembershipOracle<I, D> sulOracle, int minimalSize, int rndLength) {
+    public RandomWMethodEQOracle(MembershipOracle<I, D> sulOracle, int minimalSize, int rndLength) {
         this.sulOracle = sulOracle;
         this.minimalSize = minimalSize;
         this.rndLength = rndLength;
@@ -60,7 +59,7 @@ public class RandomWpMethod<A extends UniversalDeterministicAutomaton<?, I, ?, ?
      * @param rndLength   expected length (in addition to minimalSize) of random word
      * @param bound       specifies the bound (set to 0 for unbounded).
      */
-    public RandomWpMethod(MembershipOracle<I, D> sulOracle, int minimalSize, int rndLength, int bound) {
+    public RandomWMethodEQOracle(MembershipOracle<I, D> sulOracle, int minimalSize, int rndLength, int bound) {
         this.sulOracle = sulOracle;
         this.minimalSize = minimalSize;
         this.rndLength = rndLength;
@@ -92,16 +91,9 @@ public class RandomWpMethod<A extends UniversalDeterministicAutomaton<?, I, ?, ?
         // Then repeatedly from this for a random word
         ArrayList<I> arrayAlphabet = new ArrayList<>(inputs);
 
-        // Finally we test the state with a suffix, sometimes a global one, sometimes local
+        // Finally we test the state with a suffix
         ArrayList<Word<I>> globalSuffixes = new ArrayList<>();
         Automata.characterizingSet(hypothesis, inputs, globalSuffixes);
-
-        MutableMapping<S, ArrayList<Word<I>>> localSuffixSets = hypothesis.createStaticStateMapping();
-        for (S state : hypothesis.getStates()) {
-            ArrayList<Word<I>> suffixSet = new ArrayList<>();
-            Automata.stateCharacterizingSet(hypothesis, inputs, state, suffixSet);
-            localSuffixSets.put(state, suffixSet);
-        }
 
         Random rand = new Random();
         int currentBound = bound;
@@ -119,19 +111,8 @@ public class RandomWpMethod<A extends UniversalDeterministicAutomaton<?, I, ?, ?
             }
 
             // pick a random suffix for this state
-            // 50% chance for state testing, 50% chance for transition testing
-            if (rand.nextBoolean()) {
-                // global
-                if (!globalSuffixes.isEmpty()) {
-                    wb.append(globalSuffixes.get(rand.nextInt(globalSuffixes.size())));
-                }
-            } else {
-                // local
-                S state2 = hypothesis.getState(wb);
-                ArrayList<Word<I>> localSuffixes = localSuffixSets.get(state2);
-                if (!localSuffixes.isEmpty()) {
-                    wb.append(localSuffixes.get(rand.nextInt(localSuffixes.size())));
-                }
+            if (!globalSuffixes.isEmpty()) {
+                wb.append(globalSuffixes.get(rand.nextInt(globalSuffixes.size())));
             }
 
             Word<I> queryWord = wb.toWord();
